feat: move keys module into separate file
